src/datapipeline/requirements.txt
requests
beautifulsoup4
pdfplumber
tqdm
urllib3
chardet
typing_extensions
charset-normalizer
soupsieve
pdfminer.six
cryptography
google-cloud-storage==2.16.0
